arch(policy): split RuleSet + implementation of stateful ruleset
