+++ api/services/policy/src/engine/interfaces/index.ts
@@ -1,3 +1,3 @@
 export * from './RuleInterface';
 export * from './MetaInterface';
-export * from './CampaignMetadataRepositoryProviderInterface';
+export * from './MetadataProviderInterface';
