+++ api/services/policy/src/engine/rules/AbstractStatefulRule.ts
@@ -1,9 +1,14 @@
 import { LOWEST, priority } from '../helpers/priority';
 import { type, STATEFUL } from '../helpers/type';
-import { RuleHandlerParamsInterface, StatefulRuleInterface } from '../interfaces';
-import { MetaInterface } from '../../interfaces';
+import { ResultInterface, StatefulRuleInterface, RuleHandlerContextInterface } from '../interfaces';
+import { MetaInterface } from '../interfaces';
 
-export abstract class AbstractStatefulRule<P = any, S = any> implements StatefulRuleInterface<S> {
+interface StatefulParametersDefaultInterface {
+  uuid: string;
+}
+
+export abstract class AbstractStatefulRule<P extends StatefulParametersDefaultInterface>
+  implements StatefulRuleInterface {
   static readonly type: type = STATEFUL;
   static readonly priority: priority = LOWEST;
 
@@ -11,11 +16,13 @@ export abstract class AbstractStatefulRule<P = any, S = any> implements Stateful
   static readonly schema?: { [k: string]: any };
   static readonly description?: string;
 
-  constructor(protected parameters?: P) {}
+  constructor(protected parameters: P) {}
 
-  abstract async getState(context: RuleHandlerParamsInterface, metaGetter: MetaInterface): Promise<S>;
+  get uuid() {
+    return this.parameters.uuid;
+  }
 
-  abstract apply(context: RuleHandlerParamsInterface, state: S): void;
-
-  abstract async setState(context: RuleHandlerParamsInterface, metaGetter: MetaInterface, state: S): Promise<void>;
+  abstract getState(context: RuleHandlerContextInterface, metaGetter: MetaInterface): number;
+  abstract apply(result: ResultInterface, state: number): ResultInterface;
+  abstract setState(result: ResultInterface, state: number): number;
 }
